requirements.txt
fastapi
grpcio
grpcio-tools
aiohttp